Andrew

+++ Android/app/src/main/java/com/solution/citylogia/models/Review.kt
@@ -5,7 +5,7 @@ import java.time.OffsetDateTime
 data class Review(
         val id: Long = 0,
         val body: String = "",
-        val mark: Long = 0,
+        val mark: Float = 0.0F,
         val published_at: OffsetDateTime = OffsetDateTime.now(),
         val author: User = User(),
 )
